+++ R/helper.R
@@ -6,40 +6,20 @@ hashes = function(x) {
   map_chr(unname(x), "hash")
 }
 
-# updating join:
-# replaces values in x with values in y
-ujoin = function(x, y, key) {
-  cn = setdiff(intersect(names(x), names(y)), key)
-  expr = parse(text = paste0("`:=`(", paste0(sprintf("%1$s=i.%1$s", cn), collapse = ","), ")"))
-  x[y, eval(expr), on = key]
-}
-
-distinct = function(x, drop = TRUE) {
-  if (is.logical(x) && !drop) {
-    lvls = c(FALSE, TRUE)
-  } else if (is.factor(x)) {
-    lvls = as.character(levels(x))
-    if (drop) {
-      lvls = lvls[lvls %in% x]
-    }
-  } else {
-    lvls = unique(x)
-    lvls = lvls[!is.na(lvls)]
-  }
-  lvls
-}
-
-filter_oob_index = function(x, lower, upper) {
-  x = assert_integerish(x, coerce = TRUE)
-  x[!is.na(x) & x >= lower & x <= upper]
-}
-
 hash = function(...) {
   digest::digest(list(...), algo = "xxhash64")
 }
 
 hash_resample_iteration = function(task, learner, resampling) {
   hash(task$hash, learner$hash, resampling$hash)
+}
+
+# updating join:
+# replaces values in x with values in y
+ujoin = function(x, y, key) {
+  cn = setdiff(intersect(names(x), names(y)), key)
+  expr = parse(text = paste0("`:=`(", paste0(sprintf("%1$s=i.%1$s", cn), collapse = ","), ")"))
+  x[y, eval(expr), on = key]
 }
 
 translate_types = function(x) {
@@ -53,12 +33,3 @@ allow_partial_matching = list(
   warnPartialMatchAttr = FALSE,
   warnPartialMatchDollar = FALSE
 )
-
-# TODO: move to mlr3misc
-set_col_names = function (x, nm = x, ...) {
-    if (is.function(nm)) {
-      nm = map_chr(names2(x), nm)
-    }
-    colnames(x) = nm
-    x
-}