+++ tests/testthat/test_DataBackendCbind.R
@@ -14,7 +14,7 @@ test_that("DataBackendCbind", {
   expect_set_equal(b$rownames, 1:150)
   expect_set_equal(b$colnames, names(data))
   expect_data_table(b$data(b$rownames, b$colnames), nrows = 150, ncols = 6)
-  expect_set_equal(b$distinct(b$rownames, "Species")$Species, distinct(iris$Species, drop = FALSE))
+  expect_set_equal(b$distinct(b$rownames, "Species")$Species, distinct_values(iris$Species, drop = FALSE))
 
   x = b$missings(b$rownames, c("Petal.Width", "Petal.Length"))
   expect_equal(x, set_names(c(0L, 30L), c("Petal.Width", "Petal.Length")))