+++ tests/testthat/test_mlr_generators.R
@@ -8,7 +8,7 @@ test_that("mlr_generators", {
   for (key in keys) {
     gen = mlr_generators$get(key)
     expect_r6(gen, "Generator", private = ".generate")
-    expect_choice(gen$task_type, mlr_reflections$task_types)
+    expect_choice(gen$task_type, mlr_reflections$task_types$type)
     expect_function(gen$generate, args = "n")
     expect_class(gen$param_set, "ParamSet")
     expect_list(gen$param_set$values, names = "unique")